# server/docs/openapi.yaml
openapi: 3.0.0
info:
  description: REST API for the NDHU campus second-hand textbook marketplace.
  title: NDHU Second-Hand Book Store API
  termsOfService: http://swagger.io/terms/
  contact:
    name: SHBS Dev Team
  license:
    name: MIT
  version: "1.0"
paths:
...
  /orders:
    get:
      security:
        - BearerAuth: []
      description: Returns all orders for the authenticated user, newest first
      tags:
        - Orders
      summary: List orders
      operationId: getOrders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/model.OrderResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
    post:
      security:
        - BearerAuth: []
      description: Creates an order from the authenticated user's cart, marks listings
        as sold, and clears the cart
      tags:
        - Orders
      summary: Checkout
      operationId: checkout
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerMessageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
  /users/me:
    get:
      security:
        - BearerAuth: []
      description: Returns the authenticated user's profile information
      tags:
        - Users
      summary: Get current user
      operationId: getMe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
    put:
      security:
        - BearerAuth: []
      description: Updates the authenticated user's name and/or avatar image
      tags:
        - Users
      summary: Update current user
      operationId: updateMe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/model.SwaggerUpdateUserRequest"
        description: Fields to update
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.User"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
servers:
  - url: /api/v1
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    model.ListingWithImages:
      type: object
      properties:
        ai_confidence:
          type: number
        author:
          type: string
        condition:
          type: string
        course_code:
          type: string
        created_at:
          type: string
        department:
          type: string
        description:
          type: string
        id:
          type: string
        image_urls:
          type: array
          items:
            type: string
        isbn:
          type: string
        price:
          type: number
        seller_avatar:
          type: string
        seller_id:
          type: string
        seller_name:
          type: string
        status:
          type: string
        title:
          type: string
        updated_at:
          type: string
...
    model.OrderItemResponse:
      type: object
      properties:
        id:
          type: string
        listing:
          $ref: "#/components/schemas/model.ListingWithImages"
        price_at_purchase:
          type: number
    model.OrderResponse:
      type: object
      properties:
        created_at:
          type: string
        id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/model.OrderItemResponse"
        status:
          type: string
        total_amount:
          type: number
...
    model.SwaggerErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    model.SwaggerMessageResponse:
      type: object
      properties:
        message:
          type: string
...
    model.SwaggerUpdateUserRequest:
      type: object
      properties:
        avatar_image_id:
          type: string
        name:
          type: string
    model.User:
      type: object
      properties:
        avatar_image_id:
          type: string
        cas_id:
          type: string
        created_at:
          type: string
        email:
          type: string
        email_verified:
          type: boolean
        id:
          type: string
        is_admin:
          type: boolean
        name:
          type: string
        updated_at:
          type: string